Added test to not overwrite an existing file.

+++ rename.py
@@ -1,9 +1,6 @@
 import sys
 import os
 
-# path = smb: // sam-desktop.local/sam/downloadsjd/"The % 20New % 20York % 20Times % 20Best % 20Sellers % 20(Fiction)"
-
-# grab first snd second arguments
 old_folder = '/home/sam/Public/The New York Times Best Sellers (Fiction)'
 
 new_folder = '/home/sam/Public/TheNewYorkTimesBestSellers(Fiction)'
@@ -19,12 +16,12 @@ for filename in os.listdir(old_folder):
     data = file.read()
     file.close()
     clean_name = filename.split('. ', 1)[1]
-    print(filename)
 
-    print(clean_name)
-    new_file = open(f'{new_folder}/{clean_name}', mode='wb')
-    new_file.write(data)
-    new_file.close()
+    if not os.path.isfile(f'{new_folder}/{clean_name}'):
+        # if not os.path.exists(new_folder):
+        new_file = open(f'{new_folder}/{clean_name}', mode='wb')
+        new_file.write(data)
+        new_file.close()
 
     print(f'Done {clean_name}')
 print(" All Done")
